Change BPM to use qreal

+++ src/collection/audiocd/AudioCdMeta.h
@@ -69,7 +69,7 @@ class AudioCdTrack : public Meta::Track
 
         virtual void setTitle( const QString &newTitle );
 
-        virtual float bpm() const;
+        virtual qreal bpm() const;
 
         virtual QString comment() const;
         virtual void setComment ( const QString &newComment );
