+++ src/collection/daap/DaapMeta.h
@@ -68,7 +68,7 @@ class DaapTrack : public Meta::Track
 
         virtual void setTitle( const QString &newTitle );
 
-        virtual float bpm() const;
+        virtual qreal bpm() const;
 
         virtual QString comment() const;
         virtual void setComment ( const QString &newComment );
