+++ src/collection/proxycollection/ProxyCollectionMeta.h
@@ -54,7 +54,7 @@ namespace ProxyCollection
             Meta::YearPtr year() const;
 
             QString comment() const;
-            float   bpm() const;
+            qreal   bpm() const;
 
             double score() const;
             void setScore( double newScore );
