+++ src/core/meta/Meta.h
@@ -210,7 +210,7 @@ namespace Meta
             /** Returns the year of this track */
             virtual YearPtr year() const = 0;
             /** Returns the BPM of this track */
-            virtual float bpm() const = 0;
+            virtual qreal bpm() const = 0;
             /** Returns the comment of this track */
             virtual QString comment() const = 0;
             /** Returns the score of this track */
