+++ src/core/meta/impl/file/File.h
@@ -60,8 +60,8 @@ namespace MetaFile
 
             virtual void setTitle( const QString &newTitle );
 
-            virtual void setBpm( const float newBpm );
-            virtual float bpm() const;
+            virtual void setBpm( const qreal newBpm );
+            virtual qreal bpm() const;
 
             virtual QString comment() const;
             virtual void setComment( const QString &newComment );
