+++ src/core/meta/impl/proxy/MetaProxy_p.h
@@ -54,7 +54,7 @@ class MetaProxy::Track::Private : public QObject, public Meta::Observer
         QString cachedComposer;
         QString cachedYear;
         qint64 cachedLength;
-        float  cachedBpm;
+        qreal  cachedBpm;
         int cachedTrackNumber;
         int cachedDiscNumber;
