+++ src/core/meta/impl/stream/Stream.h
@@ -60,7 +60,7 @@ namespace MetaStream
 
             virtual void setTitle( const QString &newTitle );
 
-            virtual float bpm() const;
+            virtual qreal bpm() const;
 
             virtual QString comment() const;
             virtual void setComment( const QString &newComment );
