+++ src/services/ServiceMetaBase.h
@@ -187,7 +187,7 @@ class AMAROK_EXPORT ServiceTrack : public Meta::Track,
 
         virtual void setTitle( const QString &newTitle );
 
-        virtual float bpm() const;
+        virtual qreal bpm() const;
 
         virtual QString comment() const;
         virtual void setComment( const QString &newComment );
