+++ src/timecode/TimecodeMeta.h
@@ -71,7 +71,7 @@ public:
     virtual void setComposer( const QString &newComposer );
     virtual void setGenre( const QString &newGenre );
     virtual void setYear( const QString &newYear );
-    virtual void setBpm( const float Bpm );
+    virtual void setBpm( const qreal Bpm );
     virtual void setTitle( const QString &newTitle );
     virtual void setComment( const QString &newComment );
     virtual void setTrackNumber( int newTrackNumber );
@@ -115,8 +115,8 @@ public:
 
     virtual void setTitle( const QString &newTitle );
 
-    virtual float bpm() const;
-    virtual void setBpm ( const float newBpm );
+    virtual qreal bpm() const;
+    virtual void setBpm ( const qreal newBpm );
 
     virtual QString comment() const;
     virtual void setComment ( const QString &newComment );
@@ -176,7 +176,7 @@ private:
     qint64 m_start;
     qint64 m_end;
     qint64 m_length;
-    float m_bpm;
+    qreal m_bpm;
     int m_trackNumber;
     int m_discNumber;
     QString m_comment;
